Rename files to use UpperCamelCaseCluster/Logic for bindings
