+++ src/app/clusters/bindings/app_config_dependent_sources.cmake
@@ -26,8 +26,8 @@ TARGET_SOURCES(
     "${CLUSTER_DIR}/BindingManager.h"
     "${CLUSTER_DIR}/PendingNotificationMap.cpp"
     "${CLUSTER_DIR}/PendingNotificationMap.h"
-    "${CLUSTER_DIR}/binding-cluster.cpp"
-    "${CLUSTER_DIR}/binding-cluster.h"
+    "${CLUSTER_DIR}/BindingCluster.cpp"
+    "${CLUSTER_DIR}/BindingCluster.h"
     "${CLUSTER_DIR}/binding-table.cpp"
     "${CLUSTER_DIR}/binding-table.h"
 )
